Fix issue where truncate would bail out too early
The truncate command will do an initial scan to figure out if there are
any matching records. If this initial scan returns no records, then
dynotool used to exit, saying the table is empty. In cases where we
needed to delete relatively rare items in the table and they were not
found in the initial scan then dynotool won’t ever find them.

Fix this by allowing truncate to continue despite not finding any
matches on the initial scan.

Also:
- Improve logging so we know how mans scans we've done and how many
  items were deleted overall
- Rename a few confusing variables and fix comments